:art: Fix app env and module dependency issue
