+++ framework/src/testing/fixtures/index.ts
@@ -13,9 +13,5 @@
  *
  */
 
-export { createBlock } from './create_block';
-export { createTransaction } from './create_transaction';
-export { createGenesisBlock } from './create_genesis_block';
-export { getModuleInstance, getAccountSchemaFromModules } from './utils';
-export * from './app_env';
-export * as fixtures from './fixtures';
+export * from './genesis_block';
+export * from './accounts';
